// projects/employee_management_app/src/app/model/interface/baseModel.ts
export interface IRole {
    roleId: number,
    role: string
}

export interface IDesignation {
    designationId: number,
    designation: string
}

export interface ApiResponseModel {
    message: string,
    result: boolean,
    data: any
}

export interface Employee {
    empName: string
    empId: number
    empCode: string
    empEmailId: string
    empDesignation: string
    role: string
  }

  export interface ClientProject {
    empName: string
    empId: number
    empCode: string
    empEmailId: string
    empDesignation: string
    projectName: string
    startDate: string
    expectedEndDate: string
    clientName: string
    clientProjectId: number
  }